Changing Dim_Lead (removing duplicates on app_network join) and updating "PreProposta" rule (using ultimoUpdateEdicao column)

+++ db/ODS/views/vw_dim_lead.sql
@@ -77,7 +77,7 @@ left join lateral
   from
   	app_network an
   where
-  	l.usuario_que_indicou_id = an.user_id
+  	l.usuario_que_indicou_id::integer = an.user_id
   limit 1
 )  an
   on true;
